requirements.txt
opencv-python-headless>=4.5.0
numpy>=1.21.0
imutils>=0.5.4
pyyaml>=6.0
